Updated thrift API to use int64 for connection ID, statement ID and cursor ID
Removed Timestamp and DateTime thrift types and instead use int64 to encode both for better efficiency

Regenerated thrift files and updated ColumnValue.java.tmpl accordingly

Fixed failures in GfxdLRUDUnit (existing failures) due to increase in JVM sizes done a while back

Update the kryo reference to 4.0.0

Optimized the short/int/long read and write in OptimizedElementArray to use Platform unsafe methods
instead of bytewise manipulation

+++ gemfirexd/client/src/main/java/io/snappydata/thrift/internal/ClientPreparedStatement.java
@@ -158,7 +158,7 @@ public class ClientPreparedStatement extends ClientStatement implements
   @Override
   protected final void setCurrentRowSet(RowSet rs) {
     if (rs != null && rs.metadata != null && !rs.metadata.isEmpty()) {
-      final int stmtId = rs.getStatementId();
+      final long stmtId = rs.getStatementId();
       if (stmtId != snappydataConstants.INVALID_ID) {
         this.statementId = stmtId;
       }
@@ -962,7 +962,7 @@ public class ClientPreparedStatement extends ClientStatement implements
    * {@inheritDoc}
    */
   @Override
-  public int getStatementId() {
+  public long getStatementId() {
     return this.statementId;
   }
